django/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
psycopg2==2.8.5
pyspark==2.4.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
django-fullurl==1.1
djangorestframework==3.11.0
